extra safety: disambiguation for non-hash string names that look like hashes
Before this commit, it was impossible to create a GTA SA format file with a
string name that started with a hash and contained 8 other characters, but
wasn't actually a hash. Now, such strings can be encoded unambiguously by
adding a second hash character in front of their names. Non-hash-like strings,
like "#NAME", can still be read without modification, but will be exported as
"##NAME".